Move code back into a single file
