+++ cookbooks/aws-parallelcluster-slurm/files/default/head_node_slurm/slurm/pcluster_slurm_config_generator.py
@@ -13,20 +13,296 @@ import argparse
 import functools
 import json
 import logging
+import math
 import re
 from os import makedirs, path
 from socket import gethostname
+from typing import Dict, List, Tuple
 from urllib.parse import ParseResult, urlparse
 
 import requests
 import yaml
 from jinja2 import FileSystemLoader
 from jinja2.sandbox import SandboxedEnvironment
-from .config_renderer import QueueRenderer
 
+CONFIG_HEADER = "# This file is automatically generated by pcluster"
 METADATA_REQUEST_TIMEOUT = 60
 log = logging.getLogger()
 instance_types_data = {}
+
+# Note we have a mismatch between python path configuration for the unit tests and the integration tests that
+# prevents us to move this code in dedicated files. We will address this issue in a different task.
+# As of now we will keep these changes in the same file
+
+
+# These functions should be moved into
+# config_utils.py
+def _get_instance_types(compute_resource_config) -> List[str]:
+    """Return the InstanceTypes defined in the ComputeResource always as list."""
+    if compute_resource_config.get("Instances"):
+        return [instance.get("InstanceType") for instance in compute_resource_config["Instances"]]
+    else:
+        return [compute_resource_config["InstanceType"]]
+
+
+def _get_efa_settings(compute_resource_config) -> Tuple[bool, bool]:
+    """Return a Tuple with EFA flags."""
+    if "Efa" in compute_resource_config:
+        efa = compute_resource_config["Efa"]
+        return efa["Enabled"], efa["GdrSupport"]
+    else:
+        return False, False
+
+
+def _get_real_memory(compute_resource_config, instance_types, instance_types_info, memory_ratio) -> int:
+    """Get the RealMemory parameter to be added to the Slurm compute node configuration."""
+    schedulable_memory = compute_resource_config.get("SchedulableMemory", None)
+    if schedulable_memory is None:
+        ec2_memory = _get_min_ec2_memory(instance_types, instance_types_info)
+        return math.floor(ec2_memory * memory_ratio)
+    else:
+        return schedulable_memory
+
+
+def _get_min_ec2_memory(instance_types, instance_types_info) -> int:
+    """Return min value for EC2 memory in the instance type list."""
+    min_ec2_memory = None
+    for instance_type in instance_types:
+        instance_type_info = instance_types_info[instance_type]
+        # The instance_types_data does not have memory information for the requested instance type.
+        # In this case we set RealMemory to 1 (Slurm default value for RealMemory)
+        ec2_memory = instance_type_info.get("MemoryInfo", {}).get("SizeInMiB", 1)
+        if min_ec2_memory is None or ec2_memory < min_ec2_memory:
+            min_ec2_memory = ec2_memory
+        if min_ec2_memory == 1:
+            # ec2 memory lower bound
+            break
+    return min_ec2_memory
+
+
+def _get_min_vcpus(instance_types, instance_types_info) -> Tuple[int, int]:
+    """Return min value for vCPUs and threads per core in the instance type list."""
+    min_vcpus_count = None
+    min_threads_per_core = None
+    for instance_type in instance_types:
+        instance_type_info = instance_types_info[instance_type]
+        vcpus_info = instance_type_info.get("VCpuInfo", {})
+        # The instance_types_data does not have vCPUs information for the requested instance type.
+        # In this case we set vCPUs to 1
+        vcpus_count = vcpus_info.get("DefaultVCpus", 1)
+        if min_vcpus_count is None or vcpus_count < min_vcpus_count:
+            min_vcpus_count = vcpus_count
+        threads_per_core = vcpus_info.get("DefaultThreadsPerCore")
+        if threads_per_core is None:
+            supported_architectures = instance_type_info.get("ProcessorInfo", {}).get("SupportedArchitectures", [])
+            threads_per_core = 2 if "x86_64" in supported_architectures else 1
+        if min_threads_per_core is None or threads_per_core < min_threads_per_core:
+            min_threads_per_core = threads_per_core
+        if min_vcpus_count == 1 and min_threads_per_core == 1:
+            # vcpus and threads numbers lower bound
+            break
+    return min_vcpus_count, min_threads_per_core
+
+
+def _get_min_gpu_count_and_type(instance_types, instance_types_info, logger) -> Tuple[int, str]:
+    """Return min value for GPU and associated type in the instance type list."""
+    min_gpu_count = None
+    gpu_type_min_count = "no_gpu_type"
+    for instance_type in instance_types:
+        gpu_info = instance_types_info[instance_type].get("GpuInfo", None)
+        gpu_count = 0
+        gpu_type = "no_gpu_type"
+        if gpu_info:
+            for gpus in gpu_info.get("Gpus", []):
+                gpu_manufacturer = gpus.get("Manufacturer", "")
+                if gpu_manufacturer.upper() == "NVIDIA":
+                    gpu_count += gpus.get("Count", 0)
+                    gpu_type = gpus.get("Name").replace(" ", "").lower()
+                else:
+                    logger.info(
+                        "ParallelCluster currently does not offer native support for '%s' GPUs. "
+                        "Please make sure to use a custom AMI with the appropriate drivers in order to leverage "
+                        "GPUs functionalities",
+                        gpu_manufacturer,
+                    )
+        if min_gpu_count is None or gpu_count < min_gpu_count:
+            min_gpu_count = gpu_count
+            gpu_type_min_count = gpu_type
+        if min_gpu_count == 0:
+            # gpu number lower bound
+            break
+    return min_gpu_count, gpu_type_min_count
+
+
+# end of config_utils.py
+
+
+# These functions should be moved into
+# config_renderer.py
+class ComputeResourceRenderer:
+    """Renders a PCluster ComputeResource Config as a Slurm NodeName config, gres config or node_set element."""
+
+    def __init__(self, queue_name: str, compute_resource_config: Dict, no_gpu, memory_ratio, instance_types_info: Dict):
+        self.queue_name = queue_name
+        self.has_gpu = not no_gpu
+        self.static_nodes = compute_resource_config["MinCount"]
+        self.dynamic_nodes = compute_resource_config["MaxCount"] - self.static_nodes
+        self.name = compute_resource_config["Name"]
+        self.disable_simultaneous_multithreading = compute_resource_config["DisableSimultaneousMultithreading"]
+        self.spot_price = compute_resource_config.get("SpotPrice", None)
+        self.instance_types = _get_instance_types(compute_resource_config)
+        self.real_memory = _get_real_memory(
+            compute_resource_config, self.instance_types, instance_types_info, memory_ratio
+        )
+        self.efa_enabled, self.efa_gdr_support = _get_efa_settings(compute_resource_config)
+        self.instance_types = _get_instance_types(compute_resource_config)
+        self.vcpus_count, self.threads_per_core = _get_min_vcpus(self.instance_types, instance_types_info)
+        self.gpu_count, self.gpu_type = _get_min_gpu_count_and_type(self.instance_types, instance_types_info, log)
+
+    def static_node_name(self):
+        """Render the NodeName section for static nodes."""
+        return self._node_name("st", self.static_nodes)
+
+    def dynamic_node_name(self):
+        """Render the NodeName section for dynamic nodes."""
+        return self._node_name("dy", self.dynamic_nodes)
+
+    def vcpus(self) -> int:
+        """Return the number of vcpus according to disable_hyperthreading and instance features."""
+        return (
+            self.vcpus_count
+            if not self.disable_simultaneous_multithreading
+            else (self.vcpus_count // self.threads_per_core)
+        )
+
+    def gpus(self) -> dict:
+        """Return the number of GPUs and type for the compute resource."""
+        return {"count": self.gpu_count, "type": self.gpu_type}
+
+    def render(self):
+        """Launch the rendering process."""
+        config = ""
+        if self.static_nodes > 0:
+            config += f"NodeName={self.static_node_name()}{ self._definitions() }\n"
+        if self.dynamic_nodes > 0:
+            config += f"NodeName={self.dynamic_node_name()}{ self._definitions(dynamic=True) }\n"
+
+        return config
+
+    def as_node_set(self):
+        """Alternative rendering for the NodeSet definition."""
+        node_set = ""
+        if self.static_nodes > 0:
+            node_set += f"{self.static_node_name()},"
+        if self.dynamic_nodes > 0:
+            node_set += f"{self.dynamic_node_name()},"
+
+        return node_set
+
+    def as_gres(self):
+        """Alternative rendering for the gres config."""
+        gres_render = ""
+        if self.gpu_count > 0:
+            if self.static_nodes > 0:
+                gres_render += (
+                    f"NodeName={self.static_node_name()} Name=gpu "
+                    f"Type={ self.gpu_type } File=/dev/nvidia[0-{ self.gpu_count - 1 }]\n"
+                )
+            if self.dynamic_nodes > 0:
+                gres_render += (
+                    f"NodeName={self.dynamic_node_name()} Name=gpu "
+                    f"Type={ self.gpu_type } File=/dev/nvidia[0-{ self.gpu_count - 1 }]\n"
+                )
+
+        return gres_render
+
+    def _definitions(self, dynamic=False):
+        definitions = f" CPUs={ self.vcpus() } RealMemory={ self.real_memory } State=CLOUD {self._features(dynamic)}"
+
+        if self.has_gpu and self.gpu_count > 0:
+            definitions += f" Gres=gpu:{ self.gpu_type }:{self.gpu_count}"
+
+        return definitions
+
+    def _features(self, dynamic=False):
+        resource_type = "static"
+        if dynamic:
+            resource_type = "dynamic"
+
+        instance_type = f",{self.instance_types[0]}"
+        if len(self.instance_types) > 1:
+            # When multiple instance types are defined we do not know in advance which one will be used
+            # to launch the node. So we do not list any of them as feature
+            instance_type = ""
+
+        features = f"Feature={resource_type}{instance_type},{self.name}"
+        if self.efa_enabled:
+            features += ",efa"
+
+        if self.gpu_count > 0:
+            features += ",gpu"
+
+        return features
+
+    def _node_name(self, type, size):
+        return f"{self.queue_name}-{type}-{self.name}-[1-{ size }]"
+
+
+class QueueRenderer:
+    """Renders a PCluster Queue Config as a Slurm partition config or as gres config."""
+
+    def __init__(self, queue_config, no_gpu, memory_ratio, instance_types_info, conf_type="partition", default=False):
+        self.name = queue_config["Name"]
+        self.is_default = default
+        self.conf_type = conf_type
+        self.capacity_type = queue_config["CapacityType"]
+        self.iam = queue_config["Iam"]
+        self.compute_renderers = [
+            ComputeResourceRenderer(self.name, compute_resource_config, no_gpu, memory_ratio, instance_types_info)
+            for compute_resource_config in queue_config["ComputeResources"]
+        ]
+
+    def render(self):
+        """Launch the rendering of the required configuration."""
+        rendered = f"{CONFIG_HEADER}\n"
+        if self.conf_type == "gres":
+            rendered += self._as_gres()
+        else:
+            rendered += self._as_partition()
+
+        return rendered
+
+    def _as_partition(self):
+        partition_render = "\n"
+        for renderer in self.compute_renderers:
+            partition_render += f"{renderer.render()}"
+
+        partition_render += f"\n{self._render_node_set()}\n{self._render_partition()}\n"
+        return partition_render
+
+    def _as_gres(self):
+        gres_render = ""
+        for renderer in self.compute_renderers:
+            gres_render += f"{renderer.as_gres()}"
+
+        return gres_render
+
+    def _render_node_set(self):
+        node_set = f"NodeSet={self.name}_nodes Nodes="
+        for renderer in self.compute_renderers:
+            node_set += f"{renderer.as_node_set()}"
+
+        return node_set[:-1]
+
+    def _render_partition(self):
+        partition = f"PartitionName={self.name} Nodes={self.name}_nodes MaxTime=INFINITE State=UP"
+        if self.is_default:
+            partition += " Default=YES"
+
+        return partition
+
+
+# end of config_renderer.py
 
 
 class CriticalError(Exception):
@@ -63,7 +339,7 @@ def generate_slurm_config_files(
     output_directory = path.abspath(output_directory)
     pcluster_subdirectory = path.join(output_directory, "pcluster")
     makedirs(pcluster_subdirectory, exist_ok=True)
-    env = _get_jinja_env(template_directory, realmemory_to_ec2memory_ratio)
+    env = _get_jinja_env(template_directory)
 
     cluster_config = _load_cluster_config(input_file)
     head_node_config = _get_head_node_config()
@@ -82,7 +358,6 @@ def generate_slurm_config_files(
                 queue,
                 is_default_queue,
                 file_type,
-                env,
                 pcluster_subdirectory,
                 realmemory_to_ec2memory_ratio,
                 dryrun,
@@ -140,7 +415,6 @@ def _generate_queue_config(
     queue_config,
     is_default_queue,
     file_type,
-    jinja_env,
     output_dir,
     realmemory_to_ec2memory_ratio,
     dryrun,
@@ -196,7 +470,7 @@ def _generate_slurm_parallelcluster_configs(
         _write_rendered_template_to_file(rendered_template, filename)
 
 
-def _get_jinja_env(template_directory, realmemory_to_ec2memory_ratio):
+def _get_jinja_env(template_directory):
     """Return jinja environment with trim_blocks/lstrip_blocks set to True."""
     file_loader = FileSystemLoader(template_directory)
     # A nosec comment is appended to the following line in order to disable the B701 check.
